chore: gather chat commands service in chatController.js
